Add files via upload
A demo Software Quality & Testing project on a platform like “Digital Canvas Hub” provides artists with a modern, accessible, and global atmosphere to showcase their portfolios and sell their art securely. This system supports the visibility and creativity of artists, Also, it encourages innovation by providing tools for interaction, discovery, and transaction for which all in one unified platform. 
The documentation suggests Figma to make a prototype of “Digital Canvas Hub” at first for system interface designing. And for backend and frontend development, these languages can be used- Html, CSS, PHP, JavaScript, MySQL. For that, the software tools are- Visual studio code, XAMMP, GitHub. As for the testing for this system, Selenium will be used in terms of automated testing, for enabling tests of web functionalities and quality attributes assurance while maintaining interface and feature performance together.